Use `internal` for realz - #233 - move `*ServiceCollectionSetup` into `Microsoft.Extensions.DependencyInjection` - move `SalesforceResultCreator` and `WebHookRequestReader` into `Microsoft.AspNetCore.WebHooks`
