+++ src/Microsoft.AspNetCore.WebHooks.Receivers.AzureAlert/Extensions/AzureAlertServiceCollectionSetup.cs
@@ -3,27 +3,26 @@
 
 using System;
 using Microsoft.AspNetCore.WebHooks.Metadata;
-using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.AspNetCore.WebHooks.Internal
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
-    /// Methods to add services for the WordPress receiver.
+    /// Methods to add services for the Azure Alert receiver.
     /// </summary>
-    public static class WordPressServiceCollectionSetup
+    internal static class AzureAlertServiceCollectionSetup
     {
         /// <summary>
-        /// Add services for the WordPress receiver.
+        /// Add services for the Azure Alert receiver.
         /// </summary>
         /// <param name="services">The <see cref="IServiceCollection"/> to update.</param>
-        public static void AddWordPressServices(IServiceCollection services)
+        public static void AddAzureAlertServices(IServiceCollection services)
         {
             if (services == null)
             {
                 throw new ArgumentNullException(nameof(services));
             }
 
-            WebHookMetadata.Register<WordPressMetadata>(services);
+            WebHookMetadata.Register<AzureAlertMetadata>(services);
         }
     }
 }
