+++ src/Microsoft.AspNetCore.WebHooks.Receivers.Dropbox/Extensions/DropboxServiceCollectionSetup.cs
@@ -5,22 +5,21 @@ using System;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.WebHooks.Filters;
 using Microsoft.AspNetCore.WebHooks.Metadata;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 
-namespace Microsoft.AspNetCore.WebHooks.Internal
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
-    /// Methods to add services for the GitHub receiver.
+    /// Methods to add services for the Dropbox receiver.
     /// </summary>
-    public static class GitHubServiceCollectionSetup
+    internal static class DropboxServiceCollectionSetup
     {
         /// <summary>
-        /// Add services for the GitHub receiver.
+        /// Add services for the Dropbox receiver.
         /// </summary>
         /// <param name="services">The <see cref="IServiceCollection"/> to update.</param>
-        public static void AddGitHubServices(IServiceCollection services)
+        public static void AddDropboxServices(IServiceCollection services)
         {
             if (services == null)
             {
@@ -28,9 +27,9 @@ namespace Microsoft.AspNetCore.WebHooks.Internal
             }
 
             services.TryAddEnumerable(ServiceDescriptor.Transient<IConfigureOptions<MvcOptions>, MvcOptionsSetup>());
-            WebHookMetadata.Register<GitHubMetadata>(services);
+            WebHookMetadata.Register<DropboxMetadata>(services);
 
-            services.TryAddSingleton<GitHubVerifySignatureFilter>();
+            services.TryAddSingleton<DropboxVerifySignatureFilter>();
         }
 
         private class MvcOptionsSetup : IConfigureOptions<MvcOptions>
@@ -43,7 +42,7 @@ namespace Microsoft.AspNetCore.WebHooks.Internal
                     throw new ArgumentNullException(nameof(options));
                 }
 
-                options.Filters.AddService<GitHubVerifySignatureFilter>(WebHookSecurityFilter.Order);
+                options.Filters.AddService<DropboxVerifySignatureFilter>(WebHookSecurityFilter.Order);
             }
         }
     }
