+++ src/Microsoft.AspNetCore.WebHooks.Receivers.GitHub/Extensions/GitHubServiceCollectionSetup.cs
@@ -5,22 +5,21 @@ using System;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.WebHooks.Filters;
 using Microsoft.AspNetCore.WebHooks.Metadata;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 
-namespace Microsoft.AspNetCore.WebHooks.Internal
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
-    /// Methods to add services for the Stripe receiver.
+    /// Methods to add services for the GitHub receiver.
     /// </summary>
-    public static class StripeServiceCollectionSetup
+    internal static class GitHubServiceCollectionSetup
     {
         /// <summary>
-        /// Add services for the Stripe receiver.
+        /// Add services for the GitHub receiver.
         /// </summary>
         /// <param name="services">The <see cref="IServiceCollection"/> to update.</param>
-        public static void AddStripeServices(IServiceCollection services)
+        public static void AddGitHubServices(IServiceCollection services)
         {
             if (services == null)
             {
@@ -28,11 +27,9 @@ namespace Microsoft.AspNetCore.WebHooks.Internal
             }
 
             services.TryAddEnumerable(ServiceDescriptor.Transient<IConfigureOptions<MvcOptions>, MvcOptionsSetup>());
-            WebHookMetadata.Register<StripeMetadata>(services);
+            WebHookMetadata.Register<GitHubMetadata>(services);
 
-            services.TryAddSingleton<StripeTestEventRequestFilter>();
-            services.TryAddSingleton<StripeVerifyNotificationIdFilter>();
-            services.TryAddSingleton<StripeVerifySignatureFilter>();
+            services.TryAddSingleton<GitHubVerifySignatureFilter>();
         }
 
         private class MvcOptionsSetup : IConfigureOptions<MvcOptions>
@@ -45,10 +42,7 @@ namespace Microsoft.AspNetCore.WebHooks.Internal
                     throw new ArgumentNullException(nameof(options));
                 }
 
-                var filters = options.Filters;
-                filters.AddService<StripeTestEventRequestFilter>(StripeTestEventRequestFilter.Order);
-                filters.AddService<StripeVerifyNotificationIdFilter>(StripeVerifyNotificationIdFilter.Order);
-                filters.AddService<StripeVerifySignatureFilter>(WebHookSecurityFilter.Order);
+                options.Filters.AddService<GitHubVerifySignatureFilter>(WebHookSecurityFilter.Order);
             }
         }
     }
