+++ src/Microsoft.AspNetCore.WebHooks.Receivers.Pusher/Extensions/PusherServiceCollectionSetup.cs
@@ -5,16 +5,15 @@ using System;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.WebHooks.Filters;
 using Microsoft.AspNetCore.WebHooks.Metadata;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 
-namespace Microsoft.AspNetCore.WebHooks.Internal
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
     /// Methods to add services for the Pusher receiver.
     /// </summary>
-    public static class PusherServiceCollectionSetup
+    internal static class PusherServiceCollectionSetup
     {
         /// <summary>
         /// Add services for the Pusher receiver.
