+++ src/Microsoft.AspNetCore.WebHooks.Receivers.Salesforce/Extensions/SalesforceServiceCollectionSetup.cs
@@ -4,20 +4,20 @@
 using System;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
+using Microsoft.AspNetCore.WebHooks;
 using Microsoft.AspNetCore.WebHooks.Filters;
 using Microsoft.AspNetCore.WebHooks.Metadata;
 using Microsoft.AspNetCore.WebHooks.ModelBinding;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Microsoft.AspNetCore.WebHooks.Internal
+namespace Microsoft.Extensions.DependencyInjection
 {
     /// <summary>
     /// Methods to add services for the Salesforce receiver.
     /// </summary>
-    public static class SalesforceServiceCollectionSetup
+    internal static class SalesforceServiceCollectionSetup
     {
         /// <summary>
         /// Add services for the Salesforce receiver.
